requirements.txt

streamlit>=1.35.0
python-dotenv>=1.0.0

google-api-python-client>=2.131.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0


PyMuPDF>=1.24.0        
pdfplumber>=0.11.0,<0.12.0    

tiktoken>=0.7.0


sentence-transformers>=3.0.0
torch>=2.2.0            # CPU-only wheel is fine
transformers>=4.41.0


faiss-cpu>=1.8.0

groq>=0.9.0           
requests>=2.31.0       



numpy>=1.26.0


gitpython>=3.1.40
beautifulsoup4>=4.12.0
python-docx>=1.1.0
